Refs #43 Added support for specifying files to ignore per directory
There is a new "ignore" configuration parameter recognized in the
'.yawg.yml' file per directory. Its value is expected to be a list of
strings represeting regexp patterns.

Files whose base names match any of the regexp patterns specified with
the "ignore" parameter will not be processed during the bake. Those
files will just be ignored.