Refs #44 Added support for specifying set of files to bake per directory
There is a new "includeOnly" configuration parameter recognized in the
'.yawg.yml' file per directory. Its value is expected to be a list of
strings representing regexp patterns.

Files whose base names match any of the regexp patterns specified with
the "includeOnly" parameter will be processed during the bake. No
other files will be baked. The "includeOnly" paramater thus define the
strict list of files to be baked in that directory.

The "includeOnly" parameter is not cascaded to sub-directories.