Refs #183 Added support for baking additional dirs outside root source dir
There is a new dir baker configuration parameter called
`extraDirsHere`. This is a list of additional directories to be baked
into the current target directory. The elements of the list are file
system paths of directories. If a path is relative it is taken as
being relative to the directory currently being baked.